mcp_servers/weather_mcp/requirements.txt
flask
requests>=2.28.0
fastmcp>=0.1.0
openmeteo-requests
requests-cache
retry-requests
numpy
pandas